remove Chq dependency
The API implementation does now not rely on any OAuth or HTTP
implementation. This allows injecting the OAuth lib of our choice and
apply custom caching on the HTTP layer. The solution now contains an
extra assembly which provides a ready to use ServiceProvider that uses
Chq internal for OAuth.